third_party/blink/web_tests/http/tests/devtools/tracing/timeline-time/timeline-time-expected.txt

Test nesting of time/timeEnd records on Timeline


Running: testSimpleConsoleTime

Running: testNestedConsoleTime

Running: testUnbalancedConsoleTime

Running: testConsoleTimeWithoutConsoleTimeEnd
--------> Timestamp: Foo
--------> Timestamp: Bar
--------> Timestamp: Baz
